More collision detection fixes
- Switched to f64 for higher precision
- No more fudge (hopefully)
- Fixed bucket checking

+++ src/hlt/state.rs
@@ -1,5 +1,5 @@
 use fnv::FnvHashMap;
-use std::f32;
+use std::f64;
 use std::io::stdin;
 use hlt::parse::*;
 use hlt::strategy::Plan;
@@ -8,7 +8,7 @@ use hlt::constants::DOCK_RADIUS;
 use hlt::collision::*;
 
 pub type ID = usize;
-pub type Point = (f32, f32);
+pub type Point = (f64, f64);
 pub type Planets = FnvHashMap<ID, Planet>;
 pub type Ships = FnvHashMap<ID, Ship>;
 
@@ -26,10 +26,10 @@ pub struct Player {
 #[derive(Debug)]
 pub struct Ship {
     pub id: ID,
-    pub x: f32,
-    pub y: f32,
+    pub x: f64,
+    pub y: f64,
     pub hp: i32,
-    pub rad: f32,
+    pub rad: f64,
     pub status: Status,
     pub planet: Option<ID>,
     pub progress: i32,
@@ -55,11 +55,11 @@ impl Ship {
     }
 
     pub fn retreat_from(&self, e: &Ship, d: i32) -> Point {
-        let angle = f32::atan2(e.y - self.y, e.x - self.x);
-        (self.x - (d as f32)*angle.cos(), self.y - (d as f32)*angle.sin())
+        let angle = f64::atan2(e.y - self.y, e.x - self.x);
+        (self.x - (d as f64)*angle.cos(), self.y - (d as f64)*angle.sin())
     }
 
-    pub fn distance_to<T: ToEntity>(&self, e: &T) -> f32 {
+    pub fn distance_to<T: ToEntity>(&self, e: &T) -> f64 {
         let (x, y) = e.to_entity().pos();
         (y - self.y).hypot(x - self.x)
     }
@@ -68,10 +68,10 @@ impl Ship {
 #[derive(Debug)]
 pub struct Planet {
     pub id: ID,
-    pub x: f32,
-    pub y: f32,
+    pub x: f64,
+    pub y: f64,
     pub hp: i32,
-    pub rad: f32,
+    pub rad: f64,
     pub spots: i32,
     pub spawn: i32,
     pub owner: Option<ID>,
@@ -79,9 +79,9 @@ pub struct Planet {
 }
 
 impl Planet {
-    pub fn value(&self) -> f32 {
+    pub fn value(&self) -> f64 {
         self.owner.map_or(35.0, |_| 0.0)
-        + (self.spots as f32)
+        + (self.spots as f64)
     }
 
     pub fn is_enemy(&self, id: ID) -> bool {
@@ -102,8 +102,8 @@ impl Planet {
 
 pub struct State {
     pub id: ID,
-    pub width: f32,
-    pub height: f32,
+    pub width: f64,
+    pub height: f64,
     pub grid: Grid,
     pub players: Vec<Player>,
     pub planets: Planets,
@@ -120,8 +120,8 @@ impl State {
         stdin().read_line(&mut buffer).unwrap();
         let mut stream: Vec<_> = buffer.split_whitespace().rev().collect();
         let id = usize::take(&mut stream);
-        let width = f32::take(&mut stream);
-        let height = f32::take(&mut stream);
+        let width = f64::take(&mut stream);
+        let height = f64::take(&mut stream);
         let plan = Plan::new();
         let queue = Queue::new();
         let (players, planets, ships, grid) = take(&mut stream);
